+++ dashboard/views/request-for-quotes/view.php
@@ -70,7 +70,7 @@
                 <li class="list-stats-item list-stats-item-full">
                     <span class="label"><?php echo Labels::getLabel('LBL_ATTACHMENT', $siteLangId); ?>:</span>
                     <span class="value">
-                        <a class="link-underline text-wrap" target="blank" title="<?php echo Labels::getLabel('LBL_DOWNLOAD_FILE', $siteLangId); ?>" href="<?php echo UrlHelper::generateUrl('RequestForQuotes', 'downloadFile', array($recordId)); ?>">
+                        <a class="link-underline text-wrap" target="blank" title="<?php echo Labels::getLabel('LBL_DOWNLOAD_FILE', $siteLangId); ?>" href="<?php echo UrlHelper::generateUrl(($isSeller ? 'Seller': '') . 'RequestForQuotes', 'downloadFile', array($recordId)); ?>">
                             <svg class="svg" width="16" height="16">
                                 <use xlink:href="<?php echo CONF_WEBROOT_FRONT_URL; ?>images/retina/sprite.svg#icon-download"></use>
                             </svg>